cookbook/examples/easemytrip_voice_agent/requirements.txt
# EaseMyTrip Voice Agent Dependencies

# Core Framework
agno[all]>=2.1.0

# AI Models & APIs
openai>=1.0.0
elevenlabs>=1.0.0

# Vector Database & Embeddings
lancedb>=0.5.0

# Web Scraping & Knowledge Base
beautifulsoup4>=4.12.0
httpx>=0.27.0

# Web Interface
streamlit>=1.30.0
audio-recorder-streamlit>=0.0.8

# Server & API
fastapi>=0.109.0
uvicorn>=0.27.0

# Utilities
python-dotenv>=1.0.0
